Add list-style reading list item component
