+++ todos.md
@@ -1,7 +1,7 @@
 1. ✅ Firebase support removed 
 2. ✅ Check the sqllite to see if it stil works. (and fix) 
 3. ✅ use icons for view/edit instead of spelled out words (by default) 
-4. Make the reading list page more like a list than cards. 
+4. ✅ Make the reading list page more like a list than cards.
 5. make  the book details consistent  in each list view. 
 6. Add back some color.
